Calculate depreciation and transfer ownership figures
Calculates the depreciation by checking whether the asset was purchased
within the accounting year. If it was, the depreciation starts from the
purchase date, otherwise it starts from the accounting start date.

Likewise, if the asset is transferred to another school, the transfer
end date is used, the net book value is updated to reflect the
depreciation between the either the purchase date or the accounting
start date and the transfer date. The net book value is also reflected
in the transferred costs field, which will become the purchase cost for
the new school.